BUG FIX - changed behavior of damage multiplier to match expected. (0.8 blocks all but 20% damage applied)

+++ SticksAndStones/Models/GameComponents/Moves/StandardBlock.cs
@@ -4,7 +4,7 @@ namespace SticksAndStones.Models.GameComponents.Moves
 {
     public class StandardBlock : BaseMove
     {
-        private float _originalDefenseMultiplier;
+        private double _originalDefenseMultiplier;
         private int _runCount = 0;
         public StandardBlock(CharacterBase executioner) : base(executioner)
         {
